src/python-script/requirements.txt
ccxt==4.1.64
pandas==2.1.4
numpy==1.24.3
ta-lib==0.4.28
requests==2.31.0
python-dotenv==1.0.0